refactor variable to register object
